[Staging] Gold Beta Deployment - 1.975

+++ scripts/prep/10_ensure_resolvconf_symlink.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+if [[ $MANAGED_BY_FIREROUTER == "yes" ]]; then
+  exit
+fi
+
 if [[ -s /etc/resolv.conf && -d /run/resolvconf && ! -L /etc/resolv.conf ]]; then
   sudo cp -f /etc/resolv.conf /run/resolvconf/resolv.conf
   sudo rm /etc/resolv.conf
